cargo update

+++ src/pagination/mod.rs
@@ -56,7 +56,7 @@ use smallvec::SmallVec;
 use std::time::Duration;
 use tokio::time::sleep;
 use tokio_stream::StreamExt;
-use twilight_http::{request::channel::reaction::RequestReactionType, Error};
+use twilight_http::{error::ErrorType, request::channel::reaction::RequestReactionType};
 use twilight_model::{
     channel::{Message, Reaction, ReactionType},
     gateway::payload::ReactionAdd,
@@ -168,18 +168,22 @@ pub trait Pagination: Sync + Sized {
 
         match ctx.http.delete_all_reactions(msg.channel_id, msg.id).await {
             Ok(_) => {}
-            Err(Error::Response { status, .. }) if status.as_u16() == 403 => {
-                sleep(Duration::from_millis(100)).await;
+            Err(why) => {
+                if matches!(why.kind(), ErrorType::Response { status, .. }if status.as_u16() == 403)
+                {
+                    sleep(Duration::from_millis(100)).await;
 
-                for emote in &reactions {
-                    let request_reaction = emote.request_reaction();
+                    for emote in &reactions {
+                        let request_reaction = emote.request_reaction();
 
-                    ctx.http
-                        .delete_current_user_reaction(msg.channel_id, msg.id, request_reaction)
-                        .await?;
+                        ctx.http
+                            .delete_current_user_reaction(msg.channel_id, msg.id, request_reaction)
+                            .await?;
+                    }
+                } else {
+                    return Err(why.into());
                 }
             }
-            Err(why) => return Err(why.into()),
         }
 
         self.final_processing(ctx).await
